change int into float

+++ batch_1/input_2_numbers_print_quotient_with_decimal.py
@@ -1,6 +1,6 @@
 # ask user to input 2 numbers
-num1 = int(input("Enter the first number: "))
-num2 = int(input("Enter the second number: "))
+num1 = float(input("Enter the first number: "))
+num2 = float(input("Enter the second number: "))
 # process of getting the quotient
 quotient = num1 / num2
 # print quotient
